refactor depois de processa arquivo
