+++ content.json
@@ -1,0 +1,1 @@
+{"materiais":[{"DIG":"1493","MATERIAL":"351628*","TITULO":"VH PTV1 RETIRO78/QUATREE2","CLIENTE":"","AGENCIA":"","DUR":"17","OBSERVACAO":"PASTA PROMOÇÕES","COORDENADOR":""},{"DIG":"1495","MATERIAL":"349602","TITULO":"PARRILLA 252 30","CLIENTE":"REST PARRILLA","AGENCIA":"VISUAL COMUNICA","DUR":"30","OBSERVACAO":"SISCOM\\2019\\08\\02","COORDENADOR":""},{"DIG":"1502","MATERIAL":"337160","TITULO":"PAIS 19 HAVAN 04 DE 10","CLIENTE":"HAVAN","AGENCIA":"SOLUCAO PROPAG","DUR":"30","OBSERVACAO":"SISCOM\\2019\\08\\01","COORDENADOR":""},{"DIG":"1505","MATERIAL":"345920","TITULO":"EL BIGODON","CLIENTE":"UNIAO GONCALVES","AGENCIA":"EUREKKA COMUNIC","DUR":"30","OBSERVACAO":"SISCOM\\2019\\08\\02","COORDENADOR":""},{"DIG":"1507","MATERIAL":"347356","TITULO":"LOJAS ED MAIS B 001","CLIENTE":"EDIMAIS EMPREEN","AGENCIA":"CARLOS ROBERTO SALVUCCI - ME","DUR":"30","OBSERVACAO":"SISCOM\\2019\\07\\29","COORDENADOR":""},{"DIG":"1510","MATERIAL":"322599","TITULO":"HISTORIAS I 05 DE 05","CLIENTE":"BANCO COOPERATIVO SICREDI","AGENCIA":"MORYA SUL AGENCIA","DUR":"60","OBSERVACAO":"SISCOM\\2019\\07\\31","COORDENADOR":""},{"DIG":"1511","MATERIAL":"342758","TITULO":"SONHOS 02DE50","CLIENTE":"OTICAS CAROL","AGENCIA":"ARTPLAN PUBLIC","DUR":"15","OBSERVACAO":"SISCOM\\2019\\07\\26","COORDENADOR":""},{"DIG":"1512","MATERIAL":"347906","TITULO":"CLASSICOS 2019 D 3 DE 5","CLIENTE":"MCDONALDS","AGENCIA":"DPZ T","DUR":"15","OBSERVACAO":"SISCOM\\2019\\08\\01","COORDENADOR":""},{"DIG":"1513","MATERIAL":"349604","TITULO":"PARRILLA 252 15","CLIENTE":"REST PARRILLA","AGENCIA":"VISUAL COMUNICA","DUR":"15","OBSERVACAO":"SISCOM\\2019\\08\\01","COORDENADOR":""},{"DIG":"1514","MATERIAL":"349956","TITULO":"MANIFESTO 2019 01 DE 05","CLIENTE":"MEDLEY","AGENCIA":"PUBLICIS","DUR":"15","OBSERVACAO":"SISCOM\\2019\\07\\31","COORDENADOR":""},{"DIG":"1515","MATERIAL":"350054","TITULO":"DNA DA MARCA 45","CLIENTE":"FUND ARCO VERDE","AGENCIA":"THE HOUSE OF CREATIVITY","DUR":"45","OBSERVACAO":"SISCOM\\2019\\07\\31","COORDENADOR":""},{"DIG":"1516","MATERIAL":"350505","TITULO":"COMPARATIVO","CLIENTE":"BANCO SAFRA S A","AGENCIA":"OGILVY","DUR":"30","OBSERVACAO":"SISCOM\\2019\\08\\01","COORDENADOR":""},{"DIG":"1522","MATERIAL":"352275","TITULO":"PROMOCIONAL 6117 37 50","CLIENTE":"VIA VAREJO","AGENCIA":"YOUNG LTDA","DUR":"30","OBSERVACAO":"SISCOM\\2019\\08\\02","COORDENADOR":""},{"DIG":"1523","MATERIAL":"352604","TITULO":"ESTREIA VR VS01 - ESTREIA","CLIENTE":"ESPERANCA PRODUCOES","AGENCIA":"","DUR":"10","OBSERVACAO":"SISCOM\\2019\\08\\02","COORDENADOR":""},{"DIG":"1525","MATERIAL":"352279","TITULO":"PROMOCIONAL 6118 38 50","CLIENTE":"VIA VAREJO","AGENCIA":"YOUNG LTDA","DUR":"30","OBSERVACAO":"SISCOM\\2019\\08\\02","COORDENADOR":""},{"DIG":"1526","MATERIAL":"352601","TITULO":"GABRIEL PERSIANAS 01","CLIENTE":"PERSIANAS GABRIEL DE MARICA IN","AGENCIA":"JLG PUBLICIDADE","DUR":"15","OBSERVACAO":"SISCOM\\2019\\08\\02","COORDENADOR":""},{"DIG":"1531","MATERIAL":"352668","TITULO":"MES DOS PAIS COSTA E COSTA","CLIENTE":"COSTA COSTA MAD","AGENCIA":"","DUR":"15","OBSERVACAO":"SISCOM\\2019\\08\\02","COORDENADOR":""},{"DIG":"1533","MATERIAL":"352783","TITULO":"ROYALE INV ST 007","CLIENTE":"ROYAL SUPERMERC","AGENCIA":"KC PARTICIP","DUR":"15","OBSERVACAO":"SISCOM\\2019\\08\\02","COORDENADOR":""},{"DIG":"1537","MATERIAL":"352794","TITULO":"ROYAL INVERNO 016","CLIENTE":"ROYAL SUPERMERC","AGENCIA":"KC PARTICIP","DUR":"15","OBSERVACAO":"SISCOM\\2019\\08\\02","COORDENADOR":""},{"DIG":"1539","MATERIAL":"351538","TITULO":"FESTIVAL INVERNO 17 15SEG","CLIENTE":"CEREAIS BRAMIL","AGENCIA":"SERVIR","DUR":"15","OBSERVACAO":"SISCOM\\2019\\08\\02","COORDENADOR":""},{"DIG":"229","MATERIAL":"350833","TITULO":"GICO 100 MILHOES CELSO 1","CLIENTE":"CGCOMUNICACAO","AGENCIA":"","DUR":"45","OBSERVACAO":"SISCOM\\2019\\08\\01","COORDENADOR":""},{"DIG":"231","MATERIAL":"351041","TITULO":"GICO 100 MILHOES XOTE","CLIENTE":"CGCOMUNICACAO","AGENCIA":"","DUR":"45","OBSERVACAO":"SISCOM\\2019\\08\\01","COORDENADOR":""},{"DIG":"232","MATERIAL":"351049","TITULO":"GICO 100 MILHOES RAP","CLIENTE":"CGCOMUNICACAO","AGENCIA":"","DUR":"45","OBSERVACAO":"SISCOM\\2019\\08\\01","COORDENADOR":""},{"DIG":"233","MATERIAL":"351058","TITULO":"GICO 100 MILHOES ROCK","CLIENTE":"CGCOMUNICACAO","AGENCIA":"","DUR":"45","OBSERVACAO":"SISCOM\\2019\\08\\01","COORDENADOR":""},{"DIG":"236","MATERIAL":"348534","TITULO":"SEM CERIMO VOLTA REDONDARJ","CLIENTE":"CAPITAL HUMANO - ELENCO","AGENCIA":"","DUR":"15","OBSERVACAO":"SISCOM\\2019\\07\\30","COORDENADOR":""},{"DIG":"1494","MATERIAL":"349345","TITULO":"LIQUIDACAO 2019","CLIENTE":"OTICAS VENEZA","AGENCIA":"VENTURA PROPAG","DUR":"30","OBSERVACAO":"","COORDENADOR":""},{"DIG":"1501","MATERIAL":"299680","TITULO":"MAQUIADOR 01 DE 05","CLIENTE":"RECKITT BENCKIS","AGENCIA":"MULLEN LOWE BRASIL PUBLICIDADE","DUR":"15","OBSERVACAO":"","COORDENADOR":""},{"DIG":"1504","MATERIAL":"341500","TITULO":"PAN CONSIGNADO","CLIENTE":"BCO PAN S.A.","AGENCIA":"","DUR":"30","OBSERVACAO":"","COORDENADOR":""},{"DIG":"1524","MATERIAL":"352694","TITULO":"CH PLUGUE PATRO 03 08","CLIENTE":"CINE SHOW","AGENCIA":"ART MIDIA VR","DUR":"30","OBSERVACAO":"","COORDENADOR":""},{"DIG":"1527","MATERIAL":"352648","TITULO":"PROMOCIONAL 6112 33 50","CLIENTE":"VIA VAREJO","AGENCIA":"YOUNG LTDA","DUR":"30","OBSERVACAO":"","COORDENADOR":""},{"DIG":"1528","MATERIAL":"352656","TITULO":"PROMOCIONAL 6113 34 50","CLIENTE":"VIA VAREJO","AGENCIA":"YOUNG LTDA","DUR":"30","OBSERVACAO":"","COORDENADOR":""},{"DIG":"1529","MATERIAL":"352657","TITULO":"PROMOCIONAL 6114 35 50","CLIENTE":"VIA VAREJO","AGENCIA":"YOUNG LTDA","DUR":"30","OBSERVACAO":"","COORDENADOR":""},{"DIG":"1538","MATERIAL":"352860","TITULO":"CH REVISTA PATRO 03 08","CLIENTE":"RESENDE SHOP","AGENCIA":"SOMA COMUNICACAO","DUR":"30","OBSERVACAO":"PASTA PROMOÇÕES","COORDENADOR":""},{"DIG":"1541","MATERIAL":"352885","TITULO":"ENERGIA PARA A VIDA 01 05","CLIENTE":"ELETRONUCLEAR","AGENCIA":"LUME COMUNIC","DUR":"30","OBSERVACAO":"","COORDENADOR":""},{"DIG":"1542","MATERIAL":"352944","TITULO":"FLV 19","CLIENTE":"ESAL","AGENCIA":"EUREKKA COMUNIC","DUR":"30","OBSERVACAO":"","COORDENADOR":""},{"DIG":"238","MATERIAL":"352986","TITULO":"CH VEM AI DENTE 2019 B","CLIENTE":"TV RIO SUL","AGENCIA":"DIRETO","DUR":"45","OBSERVACAO":"PASTA PROMOÇÕES","COORDENADOR":""}]}
