+++ src/robot/processaArquivo.js
@@ -1,13 +1,11 @@
 const importExcelOpecToJson = require('./importExcelOpecToJson');
 const processaInformacao = require('./processaInformacao');
-const { moverArquivo } = require('./manipulaArquivo');
+
 const Paths = require('../paths');
 
 async function processaArquivo(path) {
-    //await importExcelOpecToJson(path);
-   
-    //await moverArquivo(path, Paths.PATH_ARQUIVO_EXCEL_PROCESSADO)
-
+    await importExcelOpecToJson(path);
     await processaInformacao();
+    
 }
 module.exports = processaArquivo
