+++ src/util/State.js
@@ -1,20 +1,20 @@
-const { escreverArquivoUtf8, lerArquivoUtf8 } = require('./manipulaArquivo');
-const contentFilePath = './content.json'
+const { escreverArquivo, lerArquivo } = require('./manipulaArquivo');
+const contentFilePath = './content.json';
  
 async function save(content){
     const contentString = JSON.stringify(content)
     //return await fs.writeFileSync(contentFilePath,contentString)   
-    return await escreverArquivoUtf8(contentFilePath,contentString);  
+    return await escreverArquivo(contentFilePath,contentString,'utf8');  
 }
 
 async function load(){
     try{
-        const fileBuffer = await lerArquivoUtf8(contentFilePath)
+        const fileBuffer = await lerArquivo(contentFilePath,'utf8')
         const contentJson = JSON.parse(fileBuffer)
         return contentJson
     }catch(err){
-        console.log(err);
-        escreverArquivoUtf8(contentFilePath,'{}');
+        //console.log(err);
+        await escreverArquivo(contentFilePath,'{}','utf8');
         return {}
    }
 }
